refactor chicken, rename css, navlink highlight
