+++ css/portafolio.css
@@ -76,6 +76,10 @@ nav {
 
   & a {
     text-decoration: inherit;
+    color: var(--color2);
+    &.active {
+      color: var(--color3);
+    }
   }
   @media (max-width: 900px) {
     &:not(.open) {
@@ -182,10 +186,20 @@ CUSTOMS
   left: 0px;
   width: 20px;
   cursor: pointer;
-  transition: all 0.1s, transform 0.6s;
+  transition: all 0.1s, transform 0.6s; 
+}
+.message {
+  position: absolute;
+  font-size: 8px;
+  background: white;
+  border-radius: 3px;
+  padding: 3px;
+  border: thin solid #111;
+  transition: all 1s;
+  user-select: none;
   &.puff {
     opacity: 0;
-    transform: scale(2);
+    transform: scale(0) translate(30px, -100px);
   }
 }
 .iam {
