+++ index.html
@@ -12,7 +12,7 @@
       rel="stylesheet"
       type="text/css"
       media="screen"
-      href="css/profile.css"
+      href="css/portafolio.css"
     />
   </head>
   <body class="layer">
